``` refactor(chat): remove unused import and commented code
Remove unused os import and commented background clearing functionality
that was not being used.

BREAKING CHANGE: The /reset command has been removed as it was
duplicated with /clear functionality.
```
