--- service/conversation.go
@@ -1,477 +1,0 @@
-package service
-
-import (
-	"encoding/json"
-	"fmt"
-	"os"
-	"sort"
-	"strconv"
-	"strings"
-
-	"github.com/activebook/gllm/data"
-	openai "github.com/sashabaranov/go-openai"
-
-	//"github.com/google/generative-ai-go/genai"
-
-	"github.com/anthropics/anthropic-sdk-go"
-	"github.com/volcengine/volcengine-go-sdk/service/arkruntime/model"
-)
-
-// ConversationManager is an interface for handling conversation history
-type ConversationManager interface {
-	SetPath(title string)
-	GetPath() string
-	Load() error
-	Save() error
-	Open(title string) error
-	Clear() error
-	Push(messages ...interface{})
-	GetMessages() interface{}
-	SetMessages(messages interface{})
-}
-
-// BaseConversation holds common fields and methods for all conversation types
-type BaseConversation struct {
-	Name string
-	Path string
-}
-
-// SetPath sets the file path for saving the conversation
-func (c *BaseConversation) SetPath(title string) {
-	if title == "" {
-		c.Path = ""
-		return
-	}
-	dir := MakeUserSubDir("gllm", "convo")
-	c.Path = GetFilePath(dir, title+".json")
-
-	// Check if file exists, if not, create an empty one
-	if _, err := os.Stat(c.Path); os.IsNotExist(err) {
-		empty := []byte("[]")
-		_ = os.WriteFile(c.Path, empty, 0644)
-	}
-}
-
-func (c *BaseConversation) GetPath() string {
-	return c.Path
-}
-
-// Common validation and file operations
-func (c *BaseConversation) readFile() ([]byte, error) {
-	if c.Name == "" {
-		return nil, nil
-	}
-
-	data, err := os.ReadFile(c.Path)
-	if err != nil {
-		if os.IsNotExist(err) {
-			return nil, nil // Return empty slice if file doesn't exist
-		}
-		return nil, fmt.Errorf("failed to read conversation file '%s': %w", c.Path, err)
-	}
-
-	// If the file is empty, return an empty slice
-	if len(data) == 0 {
-		return nil, nil
-	}
-
-	// Validate the JSON format
-	if !json.Valid(data) {
-		return nil, fmt.Errorf("invalid JSON format in conversation file '%s'", c.Path)
-	}
-
-	return data, nil
-}
-
-func (c *BaseConversation) writeFile(data []byte) error {
-	if c.Name == "" {
-		return nil
-	}
-	return os.WriteFile(c.Path, data, 0644)
-}
-
-func (c *BaseConversation) Push(messages ...interface{}) {
-}
-
-func (c *BaseConversation) GetMessages() interface{} {
-	return nil
-}
-
-func (c *BaseConversation) SetMessages(messages interface{}) {
-}
-
-// Open initializes an OpenChatConversation with the provided title, resolving
-// an index to the actual conversation name if necessary. It resets the messages,
-// sanitizes the conversation name for the path, and sets the internal path accordingly.
-// Returns an error if the title cannot be resolved.
-func (c *BaseConversation) Open(title string) error {
-	// check if it's an index
-	title, err := FindConvosByIndex(title)
-	if err != nil {
-		return err
-	}
-	// If title is still empty, no convo found
-	if title == "" {
-		return nil
-	}
-	// Set the name and path
-	c.Name = title
-	sanitized := GetSanitizeTitle(c.Name)
-	c.SetPath(sanitized)
-	return nil
-}
-
-func (c *BaseConversation) Save() error {
-	return nil
-}
-
-func (c *BaseConversation) Load() error {
-	return nil
-}
-
-func (c *BaseConversation) Clear() error {
-	if c.Name == "" {
-		return nil
-	}
-	// Clear the content of the file by writing an empty string to it
-	empty := []byte("[]")
-	err := os.WriteFile(c.Path, empty, 0644)
-	if err != nil {
-		return fmt.Errorf("failed to clear file: %w", err) // Return error if the file write fails
-	}
-	return nil
-}
-
-/*
- * OpenChat Conversation
- */
-
-// OpenChatConversation manages conversations for Volcengine model
-type OpenChatConversation struct {
-	BaseConversation
-	Messages []*model.ChatCompletionMessage
-}
-
-// PushMessages adds multiple messages to the conversation
-func (c *OpenChatConversation) Push(messages ...interface{}) {
-	for _, msg := range messages {
-		switch v := msg.(type) {
-		case *model.ChatCompletionMessage:
-			c.Messages = append(c.Messages, v)
-		case []*model.ChatCompletionMessage:
-			c.Messages = append(c.Messages, v...)
-		}
-	}
-}
-
-func (c *OpenChatConversation) GetMessages() interface{} {
-	return c.Messages
-}
-
-func (c *OpenChatConversation) SetMessages(messages interface{}) {
-	if msgs, ok := messages.([]*model.ChatCompletionMessage); ok {
-		c.Messages = msgs
-	}
-}
-
-// Save persists the conversation to disk
-func (c *OpenChatConversation) Save() error {
-	if c.Name == "" || len(c.Messages) == 0 {
-		return nil
-	}
-
-	// Most major systems (including ChatGPT and Google's Gemini) discard search results between turns
-	// Always clear content for tool messages before saving to save tokens
-	// Keep the "record" of the tool call (e.g., call_id: 123, tool: google_search) but drop the "body" of the result.
-	empty := ""
-	for _, msg := range c.Messages {
-		if msg.Role == model.ChatMessageRoleTool {
-			//msg.Content = nil // or "" if Content is a string
-			msg.Content = &model.ChatCompletionMessageContent{StringValue: &empty}
-		}
-	}
-
-	data, err := json.MarshalIndent(c.Messages, "", "  ")
-	if err != nil {
-		return fmt.Errorf("failed to serialize conversation: %w", err)
-	}
-
-	return c.writeFile(data)
-}
-
-// Load retrieves the conversation from disk
-func (c *OpenChatConversation) Load() error {
-	if c.Name == "" {
-		return nil
-	}
-
-	data, err := c.readFile()
-	if err != nil || data == nil {
-		return err
-	}
-
-	err = json.Unmarshal(data, &c.Messages)
-	if err != nil {
-		return fmt.Errorf("failed to deserialize conversation: %w", err)
-	}
-
-	if len(c.Messages) > 0 {
-		msg := c.Messages[0]
-		// Check whether all nil (aka not correct format)
-		if msg.Content == nil && msg.FunctionCall == nil && msg.ReasoningContent == nil && msg.Name == nil && len(msg.ToolCalls) == 0 {
-			return fmt.Errorf("invalid conversation format: isn't a compatible format. '%s'", c.Path)
-		}
-	}
-
-	return nil
-}
-
-// Clear removes all messages from the conversation
-func (c *OpenChatConversation) Clear() error {
-	c.Messages = []*model.ChatCompletionMessage{}
-	return c.BaseConversation.Clear()
-}
-
-/*
- * OpenAI Conversation
- */
-
-// OpenAIConversation represents a conversation using OpenAI format
-type OpenAIConversation struct {
-	BaseConversation
-	Messages []openai.ChatCompletionMessage
-}
-
-// PushMessages adds multiple messages to the conversation
-func (c *OpenAIConversation) Push(messages ...interface{}) {
-	for _, msg := range messages {
-		switch v := msg.(type) {
-		case openai.ChatCompletionMessage:
-			c.Messages = append(c.Messages, v)
-		case []openai.ChatCompletionMessage:
-			c.Messages = append(c.Messages, v...)
-		}
-	}
-}
-
-func (c *OpenAIConversation) GetMessages() interface{} {
-	return c.Messages
-}
-
-func (c *OpenAIConversation) SetMessages(messages interface{}) {
-	if msgs, ok := messages.([]openai.ChatCompletionMessage); ok {
-		c.Messages = msgs
-	}
-}
-
-// Save persists the conversation to disk
-func (c *OpenAIConversation) Save() error {
-	if c.Name == "" || len(c.Messages) == 0 {
-		return nil
-	}
-
-	// Most major systems (including ChatGPT and Google's Gemini) discard search results between turns
-	// Always clear content for tool messages before saving to save tokens
-	// Keep the "record" of the tool call (e.g., call_id: 123, tool: google_search) but drop the "body" of the result.
-	empty := ""
-	for i := range c.Messages {
-		msg := &c.Messages[i]
-		if msg.Role == openai.ChatMessageRoleTool {
-			msg.Content = empty
-		}
-	}
-
-	data, err := json.MarshalIndent(c.Messages, "", "  ")
-	if err != nil {
-		return fmt.Errorf("failed to serialize conversation: %w", err)
-	}
-
-	return c.writeFile(data)
-}
-
-// Load retrieves the conversation from disk
-func (c *OpenAIConversation) Load() error {
-	if c.Name == "" {
-		return nil
-	}
-
-	// read file
-	data, err := c.readFile()
-	if err != nil || data == nil {
-		return err
-	}
-
-	// Parse messages
-
-	if err := json.Unmarshal(data, &c.Messages); err != nil {
-		// If there's an error unmarshaling, it might be an old format
-		return fmt.Errorf("failed to parse conversation file: %v", err)
-	}
-
-	if len(c.Messages) > 0 {
-		msg := c.Messages[0]
-		if msg.Content == "" && len(msg.ToolCalls) == 0 && msg.FunctionCall == nil && msg.ReasoningContent == "" {
-			return fmt.Errorf("invalid conversation format: isn't a compatible format. '%s'", c.Path)
-		}
-	}
-
-	return nil
-}
-
-// Clear removes all messages from the conversation
-func (c *OpenAIConversation) Clear() error {
-	c.Messages = []openai.ChatCompletionMessage{}
-	return c.BaseConversation.Clear()
-}
-
-/*
- * Anthropic Conversation
- */
-
-// AnthropicConversation represents a conversation using Anthropic format
-type AnthropicConversation struct {
-	BaseConversation
-	Messages []anthropic.MessageParam
-}
-
-// PushMessages adds multiple messages to the conversation
-func (c *AnthropicConversation) Push(messages ...interface{}) {
-	for _, msg := range messages {
-		switch v := msg.(type) {
-		case anthropic.MessageParam:
-			c.Messages = append(c.Messages, v)
-		case []anthropic.MessageParam:
-			c.Messages = append(c.Messages, v...)
-		}
-	}
-}
-
-func (c *AnthropicConversation) GetMessages() interface{} {
-	return c.Messages
-}
-
-func (c *AnthropicConversation) SetMessages(messages interface{}) {
-	if msgs, ok := messages.([]anthropic.MessageParam); ok {
-		c.Messages = msgs
-	}
-}
-
-// Save persists the conversation to disk
-func (c *AnthropicConversation) Save() error {
-	if c.Name == "" || len(c.Messages) == 0 {
-		return nil
-	}
-
-	// For Anthropic, we also want to clear possibly large tool results if we want to save space,
-	// but the SDK structure is different. For now, let's just save.
-	// Optimizing storage can be a future task if needed.
-
-	data, err := json.MarshalIndent(c.Messages, "", "  ")
-	if err != nil {
-		return fmt.Errorf("failed to serialize conversation: %w", err)
-	}
-
-	return c.writeFile(data)
-}
-
-// Load retrieves the conversation from disk
-func (c *AnthropicConversation) Load() error {
-	if c.Name == "" {
-		return nil
-	}
-
-	// read file
-	data, err := c.readFile()
-	if err != nil || data == nil {
-		return err
-	}
-
-	// Parse messages
-	if err := json.Unmarshal(data, &c.Messages); err != nil {
-		return fmt.Errorf("failed to parse conversation file: %v", err)
-	}
-
-	return nil
-}
-
-// Clear removes all messages from the conversation
-func (c *AnthropicConversation) Clear() error {
-	c.Messages = []anthropic.MessageParam{}
-	return c.BaseConversation.Clear()
-}
-
-/*
- * Get the sorted list of conversations in the given directory
- * sorted by modTime descending
- */
-
-type ConvoMeta struct {
-	Name     string
-	Provider string
-	ModTime  int64
-}
-
-func GetConvoDir() string {
-	dir := data.GetConvoDirPath()
-	os.MkdirAll(dir, 0750)
-	return dir
-}
-
-func FindConvosByIndex(idx string) (string, error) {
-	if strings.TrimSpace(idx) == "" {
-		return "", nil
-	}
-	// check if it's an index
-	index, err := strconv.Atoi(idx)
-	if err == nil {
-		// It's an index, resolve to conversation name using your sorted list logic
-		convos, err := ListSortedConvos(GetConvoDir())
-		if err != nil {
-			return "", err
-		}
-		if index < 1 || index > len(convos) {
-			// handle out of range
-			return "", fmt.Errorf("conversation index out of range: %d", index)
-		} else {
-			title := convos[index-1].Name
-			return title, nil
-		}
-	} else {
-		// idx is not a index
-		return idx, nil
-	}
-}
-
-// listSortedConvos returns a slice of convoMeta sorted by modTime descending
-func ListSortedConvos(convoDir string) ([]ConvoMeta, error) {
-	files, err := os.ReadDir(convoDir)
-	if err != nil {
-		return nil, fmt.Errorf("fail to read conversation directory: %v", err)
-	}
-	var convos []ConvoMeta
-	for _, file := range files {
-		if !file.IsDir() && strings.HasSuffix(file.Name(), ".json") {
-			title := strings.TrimSuffix(file.Name(), ".json")
-			fullPath := GetFilePath(convoDir, file.Name())
-			var provider string
-			data, err := os.ReadFile(fullPath)
-			if err == nil {
-				provider = DetectMessageProvider(data)
-			}
-			info, err := os.Stat(fullPath)
-			var modTime int64
-			if err == nil {
-				modTime = info.ModTime().Unix()
-			}
-			convos = append(convos, ConvoMeta{
-				Name:     title,
-				Provider: provider,
-				ModTime:  modTime,
-			})
-		}
-	}
-	sort.Slice(convos, func(i, j int) bool {
-		return convos[i].ModTime > convos[j].ModTime
-	})
-	return convos, nil
-}
